requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-swagger==1.0.15
aiopg==1.3.1
async-timeout==3.0.1
attrs==21.2.0
cchardet==2.1.7
certifi==2020.6.20
cffi==1.14.6
chardet==4.0.0
greenlet==1.1.2
idna==3.2
MarkupSafe==1.1.1
multidict==5.2.0
psycopg2-binary==2.9.1
pycares==4.0.0
pycparser==2.20
PyYAML==5.4.1
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
yarl==1.6.3
